refactor `ApplyPendingBalanceAccountInfo` to token-2022
