+++ token/client/src/lib.rs
@@ -1,6 +1,5 @@
 #![allow(clippy::integer_arithmetic)]
 pub mod client;
-pub mod extension_info;
 pub mod output;
 pub mod token;
